Move user profiles to profile package
