+++ plugin/src/main/java/net/dzikoysk/funnyguilds/user/profile/NPCUserProfile.java
@@ -1,4 +1,4 @@
-package net.dzikoysk.funnyguilds.user;
+package net.dzikoysk.funnyguilds.user.profile;
 
 public class NPCUserProfile extends FakeUserProfile implements UserProfile {
     public NPCUserProfile() {
